(FACT-1408) Use prtdiag for productname on Solaris
Prior to this commit, facter was using prtconf to resolve the
productname and manufacturer facts on Solaris SPARC machines. We
made this change because it was believed that prtdiag was not present
on all SPARC boxes, but as far as we can tell that doesn't seem to be
the case.

Using prtconf gives a less specific fact value and doesn't work on
some Solaris machines, so we are falling back to using prtdiag since
that is what was used in Facter 2 and seems to produce better results.